db-tool: Add validation of Epoch table entries
Choose random Epoch table entry and compare the calulcated entry
to the actual table entry. This will catch and entries that were
only partially updated.
