+++ cardano-db/src/Cardano/Db/Error.hs
@@ -20,8 +20,10 @@ data LookupFail
   | DbLookupMessage !Text
   | DbLookupTxHash !ByteString
   | DbLookupTxOutPair !ByteString !Word16
+  | DbLookupEpochNo !Word64
   | DbMetaEmpty
   | DbMetaMultipleRows
+  deriving (Eq, Show)
 
 renderLookupFail :: LookupFail -> Text
 renderLookupFail lf =
@@ -32,6 +34,8 @@ renderLookupFail lf =
     DbLookupTxHash h -> "tx hash " <> base16encode h
     DbLookupTxOutPair h i ->
         Text.concat [ "tx out pair (", base16encode h, ", ", textShow i, ")" ]
+    DbLookupEpochNo e ->
+        Text.concat [ "epoch number ", textShow e ]
     DbMetaEmpty -> "Meta table is empty"
     DbMetaMultipleRows -> "Multiple rows in Meta table which should only contain one"
